requirements.txt
aiohttp==3.5.4
pytest-aiohttp==0.3.0